Fix type in print formatting

+++ src/compiler/codegen_c_json_printer.c
@@ -301,8 +301,8 @@ static int gen_json_printer_struct(fb_output_t *out, fb_compound_type_t *ct)
             tp = scalar_type_prefix(member->type.st);
             fprintf(
                     out->fp,
-                    "    flatcc_json_printer_%s_array_struct_field(ctx, %d, p, %"PRIu64", \"%.*s\", %ld, %d);\n",
-                    tp, index, (uint64_t)member->offset, (int)sym->ident->len, sym->ident->text, sym->ident->len, member->type.len);
+                    "    flatcc_json_printer_%s_array_struct_field(ctx, %d, p, %"PRIu64", \"%.*s\", %ld, %"PRIu64");\n",
+                    tp, index, (uint64_t)member->offset, (int)sym->ident->len, sym->ident->text, sym->ident->len, (uint64_t)member->type.len);
             break;
         case vt_fixed_array_compound_type_ref:
             fb_compound_name(member->type.ct, &snref);
@@ -311,8 +311,8 @@ static int gen_json_printer_struct(fb_output_t *out, fb_compound_type_t *ct)
 #if FLATCC_JSON_PRINT_MAP_ENUMS
                 tp = scalar_type_prefix(member->type.ct->type.st);
                 fprintf(out->fp,
-                        "    flatcc_json_printer_%s_enum_array_struct_field(ctx, %d, p, %"PRIu64", \"%.*s\", %ld, %d, %s_print_json_enum);\n",
-                        tp, index, (uint64_t)member->offset, (int)sym->ident->len, sym->ident->text, sym->ident->len, member->type.len, snref.text);
+                        "    flatcc_json_printer_%s_enum_array_struct_field(ctx, %d, p, %"PRIu64", \"%.*s\", %ld, %"PRIu64", %s_print_json_enum);\n",
+                        tp, index, (uint64_t)member->offset, (int)sym->ident->len, sym->ident->text, sym->ident->len, (uint64_t)member->type.len, snref.text);
                 break;
 #else
                 tp = scalar_type_prefix(member->type.ct->type.st);
